the handle adapter optimization
